Reorganize test structure with centralized mocks in testlib
- Created a dedicated testlib directory for shared test components
- Centralized CoreAudio mocks in a single implementation
- Added shared test fixtures and utilities
- Updated all test files to use the centralized mocks
- Updated README.md with documentation about the new structure
