+++ tests/README.md
@@ -78,10 +78,26 @@ The tests use different mocking strategies:
 
 4. **RenderCallback Simulation**: The fourth test file simulates the RenderCallback function to test position advancement and end-of-file detection.
 
+## Centralized Mocks Implementation
+
+All CoreAudio mocks and test fixtures are centralized in the `tests/testlib` directory:
+
+- `tests/testlib/include/coreaudio_mocks.h`: Declarations for all CoreAudio mock functions
+- `tests/testlib/src/coreaudio_mocks.cpp`: Implementations of all CoreAudio mock functions
+- `tests/testlib/include/test_utils.h`: Common test fixtures and utilities
+
+This centralized approach provides several benefits:
+
+1. **Reduced Duplication**: All mock declarations and implementations live in one place
+2. **Consistent Behavior**: All tests use the same mock implementations
+3. **Easier Maintenance**: Changes to mocks only need to be made in one place
+4. **Dependency Injection Ready**: The global `g_player` pointer is managed centrally
+
 ## Notes
 
-- The tests use preprocessor macros to replace the real CoreAudio functions with mock implementations.
+- The tests use preprocessor macros to replace the real CoreAudio functions with mock implementations, applied only in test builds via the `TESTING` definition.
 - The RenderCallback simulation test uses a global variable to access the AudioPlayer instance from the callback function.
+- The base test fixture `AudioPlayerTestBase` provides common setup/teardown and utility functions for all tests.
 
 ## Adding New Tests
